metric: standardise on float64 values
allowing it to be interface{} means a lot of risky and brittle casting.
just use float64 for everything.

+++ atc/metric/counter.go
@@ -14,7 +14,7 @@ func (m *Counter) IncDelta(delta int) {
 	atomic.AddInt64((*int64)(&m.cur), int64(delta))
 }
 
-func (m *Counter) Delta() int {
+func (m *Counter) Delta() float64 {
 	cur := atomic.SwapInt64((*int64)(&m.cur), 0)
-	return int(cur)
+	return float64(cur)
 }
